pyproject.toml
[project]
name = "mcp-legacy-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "fastmcp>=2.8.0",
    "gitpython>=3.1.44",
    "mcp-llm>=0.1.0",
    "neo4j>=5.28.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]

[tool.rye]
